split common group symbolizer stuff
