+++ Makefile
@@ -40,7 +40,7 @@ src/json/libmapnik-json.a:
 	# we first build memory intensive files with -j1
 	$(PYTHON) scons/scons.py -j1 \
 		--config=cache --implicit-cache --max-drift=1 \
-		src/renderer_common/process_group_symbolizer.os \
+		src/renderer_common/render_thunk_extractor.os \
 		src/json/libmapnik-json.a \
 		src/wkt/libmapnik-wkt.a \
 		src/css_color_grammar.os \
