+++ include/mapnik/renderer_common/process_group_symbolizer.hpp
@@ -29,6 +29,7 @@
 #include <mapnik/feature.hpp>
 #include <mapnik/feature_factory.hpp>
 #include <mapnik/renderer_common.hpp>
+#include <mapnik/renderer_common/render_thunk_extractor.hpp>
 #include <mapnik/symbolizer.hpp>
 #include <mapnik/attribute_collector.hpp>
 #include <mapnik/group/group_layout_manager.hpp>
@@ -36,175 +37,8 @@
 #include <mapnik/group/group_symbolizer_properties.hpp>
 #include <mapnik/text/glyph_positions.hpp>
 #include <mapnik/util/conversions.hpp>
-#include <mapnik/util/variant.hpp>
-#include <mapnik/label_collision_detector.hpp>
-#include <mapnik/util/noncopyable.hpp>
-#include <mapnik/svg/svg_path_adapter.hpp>
-#include <mapnik/svg/svg_path_attributes.hpp>
-
-// agg
-#include <agg_trans_affine.h>
 
 namespace mapnik {
-
-class text_symbolizer_helper;
-
-using svg::svg_path_adapter;
-using svg_attribute_type = agg::pod_bvector<svg::path_attributes>;
-
-struct virtual_renderer_common : private util::noncopyable
-{
-    virtual_renderer_common(renderer_common & common) :
-        width_(common.width_),
-        height_(common.height_),
-        scale_factor_(common.scale_factor_),
-        vars_(common.vars_),
-        shared_font_library_(common.shared_font_library_),
-        font_library_(*shared_font_library_),
-        font_manager_(common.font_manager_),
-        query_extent_(common.query_extent_),
-        t_(common.t_),
-        detector_(std::make_shared<label_collision_detector4>(common.detector_->extent())) {}
-
-    unsigned & width_;
-    unsigned & height_;
-    double & scale_factor_;
-    attributes & vars_;
-    // TODO: dirty hack for cairo renderer, figure out how to remove this
-    std::shared_ptr<font_library> & shared_font_library_;
-    font_library & font_library_;
-    face_manager_freetype & font_manager_;
-    box2d<double> & query_extent_;
-    view_transform & t_;
-    std::shared_ptr<label_collision_detector4> detector_;
-};
-
-
-// General:
-
-// The approach here is to run the normal symbolizers, but in
-// a 'virtual' blank environment where the changes that they
-// make are recorded (the detector, the render_* calls).
-//
-// The recorded boxes are then used to lay out the items and
-// the offsets from old to new positions can be used to perform
-// the actual rendering calls.
-
-// This should allow us to re-use as much as possible of the
-// existing symbolizer layout and rendering code while still
-// being able to interpose our own decisions about whether
-// a collision has occurred or not.
-
-// Thunk for rendering a particular instance of a point - this
-// stores all the arguments necessary to re-render this point
-// symbolizer at a later time.
-
-struct vector_marker_render_thunk  : util::noncopyable
-{
-    svg_path_ptr src_;
-    svg_attribute_type attrs_;
-    agg::trans_affine tr_;
-    double opacity_;
-    composite_mode_e comp_op_;
-    bool snap_to_pixels_;
-
-    vector_marker_render_thunk(svg_path_ptr const& src,
-                               svg_attribute_type const& attrs,
-                               agg::trans_affine const& marker_trans,
-                               double opacity,
-                               composite_mode_e comp_op,
-                               bool snap_to_pixels);
-
-    vector_marker_render_thunk(vector_marker_render_thunk && rhs);
-};
-
-struct raster_marker_render_thunk  : util::noncopyable
-{
-    image_rgba8 const& src_;
-    agg::trans_affine tr_;
-    double opacity_;
-    composite_mode_e comp_op_;
-    bool snap_to_pixels_;
-
-    raster_marker_render_thunk(image_rgba8 const& src,
-                               agg::trans_affine const& marker_trans,
-                               double opacity,
-                               composite_mode_e comp_op,
-                               bool snap_to_pixels);
-
-    raster_marker_render_thunk(raster_marker_render_thunk && rhs);
-};
-
-using helper_ptr = std::unique_ptr<text_symbolizer_helper>;
-
-struct text_render_thunk : util::noncopyable
-{
-    // helper is stored here in order
-    // to keep in scope the text rendering structures
-    helper_ptr helper_;
-    placements_list const& placements_;
-    double opacity_;
-    composite_mode_e comp_op_;
-    halo_rasterizer_enum halo_rasterizer_;
-
-    text_render_thunk(helper_ptr && helper,
-                      double opacity, composite_mode_e comp_op,
-                      halo_rasterizer_enum halo_rasterizer);
-
-    text_render_thunk(text_render_thunk && rhs);
-
-};
-
-// Variant type for render thunks to allow us to re-render them
-// via a static visitor later.
-
-using render_thunk = util::variant<vector_marker_render_thunk,
-                                   raster_marker_render_thunk,
-                                   text_render_thunk>;
-using render_thunk_ptr = std::unique_ptr<render_thunk>;
-using render_thunk_list = std::list<render_thunk_ptr>;
-
-// Base class for extracting the bounding boxes associated with placing
-// a symbolizer at a fake, virtual point - not real geometry.
-//
-// The bounding boxes can be used for layout, and the thunks are
-// used to re-render at locations according to the group layout.
-
-struct render_thunk_extractor
-{
-    render_thunk_extractor(box2d<double> & box,
-                           render_thunk_list & thunks,
-                           feature_impl & feature,
-                           attributes const& vars,
-                           proj_transform const& prj_trans,
-                           virtual_renderer_common & common,
-                           box2d<double> const& clipping_extent);
-
-    void operator()(markers_symbolizer const& sym) const;
-
-    void operator()(text_symbolizer const& sym) const;
-
-    void operator()(shield_symbolizer const& sym) const;
-
-    template <typename T>
-    void operator()(T const& ) const
-    {
-        // TODO: warning if unimplemented?
-    }
-
-private:
-    void extract_text_thunk(helper_ptr && helper, text_symbolizer const& sym) const;
-
-    box2d<double> & box_;
-    render_thunk_list & thunks_;
-    feature_impl & feature_;
-    attributes const& vars_;
-    proj_transform const& prj_trans_;
-    virtual_renderer_common & common_;
-    box2d<double> clipping_extent_;
-
-    void update_box() const;
-};
 
 template <typename F>
 void render_offset_placements(placements_list const& placements,
